Backend option This commit create a DatabaseBackend trait that allows to use different DB to store data, as long as there is an implementation
