+++ Cargo.toml
@@ -20,3 +20,4 @@ sqlite = "0.32.0"
 actix-rt = "2.9.0"
 tokio = "1.35.1"
 derive_more = "0.99.17"
+async-trait = "0.1.77"
