+++ src/app/navdata/airport.rs
@@ -1,7 +1,4 @@
-use crate::app::{
-    db::AppState,
-    navdata::model::{get_airport_by_icao_code, search_airport},
-};
+use crate::app::db::AppState;
 use actix_web::{get, web, HttpResponse, Responder};
 use log::{error, info};
 use serde::Deserialize;
@@ -25,14 +22,15 @@ struct FormData {
 #[get("/airport")]
 async fn airport(param: web::Query<FormData>, app_state: web::Data<AppState>) -> impl Responder {
     info!("Request received : /airport");
-    let data = search_airport(
-        param.search.clone(),
-        param.page,
-        param.country.clone(),
-        param.airport_type.clone(),
-        app_state,
-    )
-    .await;
+    let data = app_state
+        .database
+        .search_airport(
+            param.search.clone(),
+            param.page,
+            param.country.clone(),
+            param.airport_type.clone(),
+        )
+        .await;
     match data {
         Ok(data) => HttpResponse::Ok().json(json!({"status": "success", "airports" : data})),
         Err(err) => {
@@ -48,7 +46,10 @@ async fn airport_by_icao_code(
     app_state: web::Data<AppState>,
 ) -> impl Responder {
     info!("Request received : /airport/{}", icao);
-    let data = get_airport_by_icao_code(icao.to_string(), app_state).await;
+    let data = app_state
+        .database
+        .get_airport_by_icao_code(icao.to_string())
+        .await;
     match data {
         Ok(data) => HttpResponse::Ok().json(json!({"status": "success", "airport" : data})),
         Err(err) => {
