+++ src/app/navdata/mod.rs
@@ -1,3 +1,2 @@
 pub mod airport;
-pub mod model;
 pub mod navaid;
