+++ src/app/navdata/navaid.rs
@@ -1,7 +1,4 @@
-use crate::app::{
-    db::AppState,
-    navdata::model::{get_navaid_by_icao_code, search_navaid},
-};
+use crate::app::db::AppState;
 use actix_web::{get, web, HttpResponse, Responder};
 use log::{error, info};
 use serde::Deserialize;
@@ -25,14 +22,15 @@ struct FormData {
 #[get("/navaid")]
 async fn navaid(param: web::Query<FormData>, app_state: web::Data<AppState>) -> impl Responder {
     info!("Request received : /navaid");
-    let data = search_navaid(
-        param.search.clone(),
-        param.page,
-        param.country.clone(),
-        param.navaid_type.clone(),
-        app_state,
-    )
-    .await;
+    let data = app_state
+        .database
+        .search_navaid(
+            param.search.clone(),
+            param.page,
+            param.country.clone(),
+            param.navaid_type.clone(),
+        )
+        .await;
     match data {
         Ok(data) => HttpResponse::Ok().json(json!({"status": "success", "navaid" : data})),
         Err(err) => {
@@ -48,7 +46,10 @@ async fn navaid_by_icao_code(
     app_state: web::Data<AppState>,
 ) -> impl Responder {
     info!("Request received : /navaid/{}", icao);
-    let data = get_navaid_by_icao_code(icao.to_string(), app_state).await;
+    let data = app_state
+        .database
+        .get_navaid_by_icao_code(icao.to_string())
+        .await;
     match data {
         Ok(data) => HttpResponse::Ok().json(json!({"status": "success", "navaid" : data})),
         Err(err) => {
